AzureVMControl -> AzureVMControlImpl, VMControl -> AzureVMControl
