+++ azure/src/main/java/com/peterphi/std/azure/guice/AzureModule.java
@@ -7,8 +7,6 @@ import com.google.inject.Singleton;
 import com.microsoft.azure.Azure;
 import com.microsoft.azure.management.compute.VirtualMachines;
 import com.microsoft.rest.credentials.ServiceClientCredentials;
-import com.peterphi.std.azure.AzureVMControl;
-import com.peterphi.std.azure.VMControl;
 
 /**
  * Created by bmcleod on 05/09/2016.
@@ -20,7 +18,6 @@ public class AzureModule extends AbstractModule
 	{
 		bind(ServiceClientCredentials.class).toProvider(ServiceClientCredentialsProvider.class);
 		bind(Azure.class).toProvider(AzureProvider.class);
-		bind(VMControl.class).to(AzureVMControl.class).in(Singleton.class);
 	}
